Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

MultiPrecisionR2.FPMPNLPModel

MultiPrecisionR2.MPCounters

MultiPrecisionR2.MPR2Params

MultiPrecisionR2.MPR2Precisions

MultiPrecisionR2.MPR2Precisions

MultiPrecisionR2.MPR2Solver

MultiPrecisionR2.CheckMPR2ParamConditions

MultiPrecisionR2.GradIntervalEval_test

MultiPrecisionR2.GradTypeStableTest

MultiPrecisionR2.MPR2

MultiPrecisionR2.ObjIntervalEval_test

MultiPrecisionR2.ObjTypeStableTest

MultiPrecisionR2.check_overflow

MultiPrecisionR2.computeCandidate!

MultiPrecisionR2.computeModelDecrease!

MultiPrecisionR2.computeMu

MultiPrecisionR2.computeStep!

MultiPrecisionR2.compute_f_at_c_default!

MultiPrecisionR2.compute_f_at_x_default!

MultiPrecisionR2.compute_g_default!

MultiPrecisionR2.get_id

MultiPrecisionR2.gradReachPrec

MultiPrecisionR2.gradReachPrec!

MultiPrecisionR2.graderrmp

MultiPrecisionR2.graderrmp!

MultiPrecisionR2.increment!

MultiPrecisionR2.objReachPrec

MultiPrecisionR2.objerrmp

MultiPrecisionR2.recomputeMu!

MultiPrecisionR2.recomputeMuPrecSelection!

MultiPrecisionR2.recompute_g_default!

MultiPrecisionR2.reset!

MultiPrecisionR2.reset!

MultiPrecisionR2.selectPic_default!

MultiPrecisionR2.selectPif!

MultiPrecisionR2.sum_counters

MultiPrecisionR2.sum_counters

MultiPrecisionR2.umpt!

MultiPrecisionR2.update_struct!

MultiPrecisionR2.γfunc_test_error_bound

MultiPrecisionR2.γfunc_test_template

NLPModels.neval_cons

NLPModels.neval_cons_lin

NLPModels.neval_cons_nln

NLPModels.neval_grad

NLPModels.neval_hess

NLPModels.neval_hprod

NLPModels.neval_jac

NLPModels.neval_jac_lin

NLPModels.neval_jac_nln

NLPModels.neval_jcon

NLPModels.neval_jgrad

NLPModels.neval_jhess

NLPModels.neval_jhprod

NLPModels.neval_jprod

NLPModels.neval_jprod_lin

NLPModels.neval_jprod_nln

NLPModels.neval_jtprod

NLPModels.neval_jtprod_lin

NLPModels.neval_jtprod_nln

NLPModels.neval_obj

​

MultiPrecisionR2.FPMPNLPModel — Type

FPMPNLPModel(Model::AbstractNLPModel{D,S},FPList::Vector{K}; kwargs...) where {D,S,K<:DataType}
FPMPNLPModel(f,x0, FPList::Vector{DataType}; kwargs...)

Floating-Point Multi-Precision Non Linear Model (FPMPNLPModel) structure. This structure is intended to extend NLPModel structure to multi-precision.

Provides errors on objective function and grandient evaluation (see objerrmp and graderrmp).

The error models are :

ojective: |fl(f(x)) - f(x)| ≤ ωf

gradient: ||fl(∇f(x)) - ∇f(x)||₂ ≤ ||fl(∇f(x))||₂ ωg.

ωf and ωg are evaluated either using:

interval analysis (can be very slow)

based on relative error assumption (see ωfRelErr and ωgRelErr field description below)

Fields

Model::AbstractNLPModel : NLPModel

FPList::Vector{DataType} : List of floating point formats

EpsList::Vector{H} : List of machine epsilons of the floating point formats in FPList

UList::Vector{H} : List of unit round-off of the floating point formats in FPList

γfunc : callback function for dot product rounding error parameter |γ|, |fl(x.y) - x.y| ≤ |x|.|y| γ. Expected signature is γfunc(n::Int,u::H) and output is H. Default callback γfunc(n::Int,u::H) = n*u is implemented upon instantiation.

ωfRelErr::Vector{H} : List of relative error factor for objective function evaluation for formats in FPList. Error model is |f(x)-fl(f(x))| ≤ ωfRelErr * |fl(f(x))|

ωgRelErr::Vector{H} : List of relative error factor for gradient evaluation for formats in FPList. Error model is ||∇f(x)-fl(∇f(x))||₂ ≤ ωgRelErr * ||fl(∇f(x))||₂

ObjEvalMode::Int : Evalutation mode for objective and error. Set automatically upon instantiation. Possible values:

INT_ERR : interval evaluation of objective (chosen as middle of the interval) and error

REL_ERR : classical evaluation and use relative error model (with ωfRelErr value)

GradEvalMode::Int : Evalutation mode for gradient and error. Set automatically upon instantiation. Possible values:

INT_ERR : interval evaluation of gradient (chosen as middle of interval vector) and error

REL_ERR : classical evaluation and use relative error model (with ωgRelErr value)

Constructors:

FPMPModel(Model, FPList; kwargs...) :

Model::AbstractNLPModel: Base model

FPList::Vector{DataType}: List of FP formats that can be used for evaluations

FPModels(f,x0::Vector,FPList::Vector{DataType}; kwargs...) : Instanciate a ADNLPModel with f and x0 and call above constructor

f : objective function

x0 : initial solution

FPList::Vector{DataType}: List of FP formats that can be used for evaluations

Keyword arguments:

HPFormat=Float64 : high precision format (must be at least as accurate as FPList[end])

γfunc=nothing : use default if not provided (see Fields section above)

ωfRelErr=HPFormat.(sqrt.(eps.(FPList))): use relative error model by default for objective evaluation

ωgRelErr=HPFormat.(sqrt.(eps.(FPList))): use relative error model by default for gradient evaluation

obj_int_eval = false : if true, use interval arithmetic for objective value and error evaluation

grad_int_eval = false : if true, use interval arithmetic for gradient value and error evaluation

Checks upon instantiation

Some checks are performed upon instantiation. These checks include:

Length consistency of vector fields: FPList, EpsList

HPFormat is at least as accurate as the highest precision floating point format in FPList. Ideally HPFormat is more accurate to ensure the numerical stability of MPR2 algorithm.

Interval evaluations: it might happen that interval evaluation of objective function and/or gradient is type-unstable or returns an error. The constructor returns an error in this case. This type of error is most likely due to IntervalArithmetic.jl.

FPList is ordered by increasing floating point format accuracy

These checks can return @warn or error.

Examples

using MultiPrecisionR2

T = [Float16, Float32]
f(x) = x[1]^2 + x[2]^2
x = zeros(2)
mpnlp = FPMPNLPModel(f,x0,T)

using MultiPrecisionR2
using OptimizationProblems
using OptimizationProblems.ADNLPProblems
using ADNLPModels
using BFloat16s

T = [BFloat16, Float16, Float32]
nlp = woods()
mpnlp = (nlp,T)

MultiPrecisionR2.MPCounters — Type

MPCounters

Struct for storing the number of function evaluations with each floating point format. The fields are the same as NLPModels.Counters, but are Dict{DataType,Int} instead of Int.

MPCounters(FPformats::Vector{DataType})

Creates an empty MPCounters struct for types in the vector FPformats.

Example

using MultiPrecisionR2.jl
FPformats = [Float16, Float32]
cntrs = MPCounters(FPformats)

MultiPrecisionR2.MPR2Params — Type

MPR2Params(LPFormat::DataType, HPFormat::DataType)

MPR2 parameters structure.

Fields

η₀::H : controls objective function error tolerance, convergence condition is ωf ≤ η₀ ΔT (see FPMPNLPModel for details on ωf)

η₁::H : step successful if ρ ≥ η₁ (update incumbent)

η₂::H : step very successful if ρ ≥ η₂ (decrease σ ⟹ increase step length)

κₘ::H : tolerance on gradient evaluation error, μ ≤ κₘ (see computeMu)

γ₁::L : σk+1 = σk * γ₁ if ρ ≥ η₂

γ₂::L : σk+1 = σk * γ₂ if ρ < η₁

Parameters

H must correspond to MPnlp.HPFormat with MPnlp given as input of MPR2

L must correspond to MPnlp.FPList[1], i.e the lowest precision floating point format used by MPnlp given as input of MPR2

Conditions

Parameters must statisfy the following conditions:

0 ≤ η₀ ≤ 1/2*η₁

0 ≤ η₁ ≤ η₂ < 1

η₀+κₘ/2 ≤0.5*(1-η₂)

η₂<1

0<γ₁<1<γ₂

Instiates default values:

η₀::H = 0.01

η₁::H = 0.02

η₂::H = 0.95

κₘ::H = 0.02

γ₁::L = 2^(-2)

γ₂::L = 2

MultiPrecisionR2.MPR2Precisions — Type

MPR2Precisions(π::Int)

FP format index storage structure.

Stores FP formats index in FPMPNLPModel.FPList of obj, grad, model reduction and norms evaluations, and FP format index of MPR2 algorithm vector variables.

Fields

πx::Int : Current FP format of current incumbent x

πnx::Int : FP format used for x norm evaluation

πs::Int : Current FP format of steps

πns::Int : FP format used for s norm evaluation

πc::Int : Current FP format of candidate c

πf::Int : FP format used for objective evaluation at x

πf⁺::Int : FP format used for objective evaluation at c

πg::Int : FP format used for gradient evaluation at c

πΔ::Int : FP format used for model reduction computation

MultiPrecisionR2.MPR2Precisions — Method

MPR2Precisions(π::Int)

FP format index storage structure.

Stores FP formats index in FPMPNLPModel.FPList of obj, grad, model reduction and norms evaluations, and FP format index of MPR2 algorithm vector variables.

Fields

πx::Int : Current FP format of current incumbent x

πnx::Int : FP format used for x norm evaluation

πs::Int : Current FP format of steps

πns::Int : FP format used for s norm evaluation

πc::Int : Current FP format of candidate c

πf::Int : FP format used for objective evaluation at x

πf⁺::Int : FP format used for objective evaluation at c

πg::Int : FP format used for gradient evaluation at c

πΔ::Int : FP format used for model reduction computation

MultiPrecisionR2.MPR2Solver — Type

MPR2Solver(MPnlp::FPMPNLPModel)

Solver structure containing all the variables necessary to MRP2.

Fields:

x::T: incumbent

g::T : gradient

s::T : step

c::T : candidate

π::MPR2Precisions : FP format indices (precision) structure

p::MPR2Params : MPR2 parameters

x_norm::H : norm of x

s_norm::H : norm of s

g_norm::H : norm of g

ΔT::H : model decrease

ρ::H : success ratio

ϕ::H : guaranteed upper bound on ||x||/||s||

ϕhat::H : computed value of ||x||/||s||

μ::H : error indicator

f::H : objective value at x

f⁺::H : objective value at c

ωf::H : objective evaluation error at x, |f(x) - fl(f(x))| <= ωf

ωf⁺::H : objective evaluation error at c, |f(c) - fl(f(c))| <= ωf⁺

ωg::H : gradient evaluation error at c, ||∇f(c) - fl(∇f(c))||₂ <= ωg||fl(∇f(c))||₂

ωfBound::H : error tolerance on objective evaluation

σ::H : regularization parameter

πmax::Int : number of FP formats available for evaluations

init::Bool : initialized with true, set to false when entering main loop

MultiPrecisionR2.CheckMPR2ParamConditions — Method

CheckMPR2ParamConditions(p::MPR2Params{H})

Check if the MPR2 parameters conditions are satified. See MPR2Params for parameter conditions.

MultiPrecisionR2.GradIntervalEval_test — Method

GradIntervalEval_test(nlp::AbstractNLPModel,FPList::AbstractArray)

Test interval evaluation of gradient for all FP formats. Test fails and return an error if interval evaluation returns an error. See [FPMPNLPModel], [AbstractNLPModel]

MultiPrecisionR2.GradTypeStableTest — Method

GradTypeStableTest(nlp::AbstractNLPModel, FPList::AbstractArray)

Tests if objective evaluation of nlp is type stable for FP format in FPList

MultiPrecisionR2.MPR2 — Method

MPR2(MPnlp; kwargs...)

An implementation of the quadratic regularization algorithm with dynamic selection of floating point format for objective and gradient evaluation, robust against finite precision rounding errors. Type parameters are: S::AbstractVector, H::AbstractFloat, T::AbstractFloat, E::DataType

Arguments

MPnlp::FPMPNLPModel : Multi precision model, see FPMPNLPModel

Keyword agruments:

x₀::S = MPnlp.Model.meta.x0 : initial guess

par::MPR2Params = MPR2Params(MPnlp.FPList[1],H) : MPR2 parameters, see MPR2Params for details

atol::H = H(sqrt(eps(T))) : absolute tolerance on first order criterion

rtol::H = H(sqrt(eps(T))) : relative tolerance on first order criterion

max_eval::Int = -1: maximum number of evaluation of the objective function.

max_iter::Int = 1000 : maximum number of iteration allowed

σmin::T = sqrt(T(MPnlp.EpsList[end])) : minimal value for regularization parameter. Value must be representable in any of the floating point formats of MPnlp.

verbose::Int=0 : display iteration information if > 0

e::E : user defined structure, used as argument for compute_f_at_x!, compute_f_at_c! compute_g! and recompute_g! callback functions.

compute_f_at_x! : callback function to select precision and compute objective value and error bound at the current point. Allows to reevaluate the objective at x if more precision is needed.

compute_f_at_c! : callback function to select precision and compute objective value and error bound at candidate.

compute_g! : callback function to select precision and compute gradient value and error bound. Called at the end of main loop.

recompute_g! : callback function to select precision and recompute gradient value if more precision is needed. Called after step, candidate and model decrease computation in main loop.

selectPic! : callback function to select FP format of c at the next iteration

Outputs

GenericExecutionStats: execution stats containing information about algorithm execution (nb. of iteration, termination status, ...). See SolverCore.jl

Example

T = [Float32, Float64]
f(x) = sum(x.^2)
x = ones(Float32,2)
mpnlp = FPMPNLPModel(f,x,T)
MPR2(mpnlp)

MultiPrecisionR2.ObjIntervalEval_test — Method

ObjIntervalEval_test(nlp::AbstractNLPModel,FPList::AbstractArray)

Test interval evaluation of objective for all formats in FPList. Test fails and return an error if interval evaluation returns an error. See [FPMPNLPModel], [AbstractNLPModel]

MultiPrecisionR2.ObjTypeStableTest — Method

ObjTypeStableTest(nlp::AbstractNLPModel, FPList::AbstractArray)

Tests if objective evaluation of nlp is type stable for FP format in FPList.

MultiPrecisionR2.check_overflow — Method

check_overflow(f)

f::AbstractFloat: Returns true if f is inf or nan, false otherwise.

f::Interval : Returns true if diam(f) is inf or nan, false otherwise.

f::AbstractVector{AbstractFloat} : Returns true if on element of f is inf or nan, false otherwise.

f::AbstractVector{Interval}: Returns true if on element of diam(f) is inf, false otherwise.

MultiPrecisionR2.computeCandidate! — Method

computeCandidate!(c::T, x::T, s::T, FP::Vector{DataType}, π::MPR2Precisions) where {T <: Tuple}

Compute candidate with proper FP format to avoid underflow and overflow

Arguments

x::T : incumbent

s::T : step

FP::Vector{Int} : Available floating point formats

π::MPR2Precisions : FP format index storage structure

Modified arguments:

c::T : updated with candidate

π::MPR2Precisions : π.πc updated with FP format index used for computation

Outputs:

::bool : false if over/underflow occur with highest precision FP format, true otherwise

MultiPrecisionR2.computeModelDecrease! — Method

computeModelDecrease!(g::T,s::T,solver::MPR2Solver,FP::Vector{DataType},π::MPR2Precisions) where {T <: Tuple}

Compute model decrease with FP format avoiding underflow and overflow

Arguments

g::T : gradient

s::T : step

solver::MPR2Solver : solver structure, stores intermediate variables

FP::Vector{Int} : Available floating point formats

π::MPR2Precisions : FP format index storage structure

Modified Arguments

solver::MPR2Solver : solver.ΔT updated

π::MPR2Precisions : π.πΔ updated

Outputs

::bool : false if over/underflow occur with highest precision FP format, true otherwise

MultiPrecisionR2.computeMu — Method

computeMu(m::FPMPNLPModel, solver::MPR2Solver{T,H}; π::MPR2Precisions = solver.π)

Compute μ value for gradient error ωg, ratio ϕ = ||x||/||s|| and rounding error models

MultiPrecisionR2.computeStep! — Method

computeStep!(s::T, g::T, σ::H, FP::Vector{DataType}, π::MPR2Precisions) where {T <: Tuple, H}

Compute step with proper FP format to avoid underflow and overflow

Arguments

s::T : step

g::T : gradient

πg::Int : g FP index

σ::H : regularization parameter

FP::Vector{Int} : Available floating point formats

π::MPR2Precisions : FP format index storage structure

Modified arguments :

s::T : updated with computed step

π::MPR2Precisions : π.πs updated with FP format used for step computation

Outputs

::bool : false if over/underflow occurs, true otherwise

MultiPrecisionR2.compute_f_at_c_default! — Method

compute_f_at_c_default!(m::FPMPNLPModel, solver::MPR2Solver{T,H}, stats::GenericExecutionStats, e::E)

Compute objective function at the candidate. Updates related fields of solver.

Outputs:

::bool: returns false if couldn't reach sufficiently small evaluation error or overflow occured.

MultiPrecisionR2.compute_f_at_x_default! — Method

compute_f_at_x_default!(m::FPMPNLPModel, solver::MPR2Solver{T,H}, stats::GenericExecutionStats, e::E)

Compute objective function at the current incumbent. Updates related fields of solver.

Outputs:

::bool : returns false if couldn't reach sufficiently small evaluation error or overflow occured.

MultiPrecisionR2.compute_g_default! — Method

compute_g_default!(m::FPMPNLPModel, solver::MPR2Solver{T,H}, stats::GenericExecutionStats, e::E)

Compute gradient at x if solver.init == true (first gradient eval outside of main loop), at c otherwise.

Outputs:

::bool : always true (needed to comply with callback template)

MultiPrecisionR2.get_id — Method

get_id(m::FPMPNLPModel, FPFormat::DataType)

Returns the index of FPFormat in m.FPList.

MultiPrecisionR2.gradReachPrec! — Method

gradReachPrec!(m::FPMPNLPModel{H}, x::T, g::T, err_bound::H; π::Int = 1) where {T <: Tuple, H}
gradReachPrec(m::FPMPNLPModel{H}, x::T, err_bound::H; π::Int = 1) where {T <: Tuple, H}

Evaluates gradient and increase model precision to reach necessary error bound or to avoid overflow.

Arguments

m::FPMPNLPModel{H}: multi precision model

x::T: tuple containing value of x in the FP formats of FPList

g::T: gradient container

π: Initial ''guess'' precision level that can provide evaluation error lower than err_bound, use 1 by default (lowest precision)

Outputs

(g): gradient, returned only with gradReachPrec call

ωg: objective evaluation error

id: precision level used for evaluation

Modified

(g): updated with the gradient value, only with gradReachPrec! call

There is no guarantee that ωg ≤ err_bound. This case happens if the highest precision FP format is not accurate enough.

MultiPrecisionR2.gradReachPrec — Method

gradReachPrec!(m::FPMPNLPModel{H}, x::T, g::T, err_bound::H; π::Int = 1) where {T <: Tuple, H}
gradReachPrec(m::FPMPNLPModel{H}, x::T, err_bound::H; π::Int = 1) where {T <: Tuple, H}

Evaluates gradient and increase model precision to reach necessary error bound or to avoid overflow.

Arguments

m::FPMPNLPModel{H}: multi precision model

x::T: tuple containing value of x in the FP formats of FPList

g::T: gradient container

π: Initial ''guess'' precision level that can provide evaluation error lower than err_bound, use 1 by default (lowest precision)

Outputs

(g): gradient, returned only with gradReachPrec call

ωg: objective evaluation error

id: precision level used for evaluation

Modified

(g): updated with the gradient value, only with gradReachPrec! call

There is no guarantee that ωg ≤ err_bound. This case happens if the highest precision FP format is not accurate enough.

MultiPrecisionR2.graderrmp! — Method

graderrmp(m::FPMPNLPModel, x::V)
graderrmp!(m::FPMPNLPModel{H}, x::V, g::V) where {H, S, V<:AbstractVector{S}}
graderrmp!(m::FPMPNLPModel{H}, x::V, g::V, ::Val{INT_ERR}) where {H, S, V<:AbstractVector{S}}
graderrmp!(m::FPMPNLPModel{H}, x::V, g::V, ::Val{REL_ERR}) where {H, S, V<:AbstractVector{S}}

Evaluates the gradient g and the relative evaluation error ωg. The two functions with the extra argument ::Val{INT_ERR} and ::Val{REL_ERR} handles the interval and "classic" evaluation of the objective and the error, respectively.

Arguments

m::FPMPNLPModel : multi-precision model

x::V: where the gradient is evaluated

g::V: container for gradient

Outputs

g::Vector{S}: gradient value, only returned with graderrmp.

ωg <: AbstractFloat: evaluation error satisfying: ||∇f(x) - fl(∇f(x))||₂ ≤ ωg||g||₂ with fl() the floating point evaluation.

Note: ωg FP format may be different than S

Modified

g::V: updated with gradient value

Overflow cases:

Interval evaluation: if at least one element of g has infinite diameter, returns [0]ⁿ, Inf

Classical evaluation: if at least one element of g overflows, returns g, Inf

MultiPrecisionR2.graderrmp — Method

graderrmp(m::FPMPNLPModel, x::V)
graderrmp!(m::FPMPNLPModel{H}, x::V, g::V) where {H, S, V<:AbstractVector{S}}
graderrmp!(m::FPMPNLPModel{H}, x::V, g::V, ::Val{INT_ERR}) where {H, S, V<:AbstractVector{S}}
graderrmp!(m::FPMPNLPModel{H}, x::V, g::V, ::Val{REL_ERR}) where {H, S, V<:AbstractVector{S}}

Evaluates the gradient g and the relative evaluation error ωg. The two functions with the extra argument ::Val{INT_ERR} and ::Val{REL_ERR} handles the interval and "classic" evaluation of the objective and the error, respectively.

Arguments

m::FPMPNLPModel : multi-precision model

x::V: where the gradient is evaluated

g::V: container for gradient

Outputs

g::Vector{S}: gradient value, only returned with graderrmp.

ωg <: AbstractFloat: evaluation error satisfying: ||∇f(x) - fl(∇f(x))||₂ ≤ ωg||g||₂ with fl() the floating point evaluation.

Note: ωg FP format may be different than S

Modified

g::V: updated with gradient value

Overflow cases:

Interval evaluation: if at least one element of g has infinite diameter, returns [0]ⁿ, Inf

Classical evaluation: if at least one element of g overflows, returns g, Inf

MultiPrecisionR2.increment! — Method

increment!(nlp, s)

Increment counter s of problem nlp.

MultiPrecisionR2.objReachPrec — Method

objReachPrec(m::FPMPNLPModel{H}, x::T, err_bound::H; π::Int = 1) where {T <: Tuple, H}

Evaluates objective and increase model precision to reach necessary error bound or to avoid overflow.

Arguments

m::FPMPNLPModel{H}: multi precision model

x::T: tuple containing value of x in the FP formats of FPList

err_bound::H : evaluation error tolerance

π::Int: Initial ''guess'' for FP format that can provide evaluation error lower than err_bound, use 1 by default (lowest precision FP format)

Outputs

f: objective value at x

ωf: objective evaluation error

id: precision level used for evaluation

There is no guarantee that ωf ≤ err_bound, happens if highest precision FP format is not accurate enough.

MultiPrecisionR2.objerrmp — Method

objerrmp(m::FPMPNLPModel, x::AbstractVector{T})
objerrmp(m::FPMPNLPModel, x::AbstractVector{S}, ::Val{INT_ERR})
objerrmp(m::FPMPNLPModel, x::AbstractVector{S}, ::Val{REL_ERR})

Evaluates the objective and the evaluation error. The two functions with the extra argument ::Val{INT_ERR} and ::Val{REL_ERR} handles the interval and "classic" evaluation of the objective and the error, respectively.

Arguments

x::Vector{S}: where to evaluate the objective, can be either a vector of AbstractFloat or a vector of Intervals.

Outputs

fl(f(x)): finite-precision evaluation of the objective at x

ωf <: AbstractFloat: evaluation error, |f(x)-fl(f(x))| ≤ ωf with fl() the floating point evaluation.

Overflow cases:

Interval evaluation: overflow occurs if the diameter of the interval enclosing f(x) is Inf. Returns 0, Inf

Classical evaluation:

If obj(x) = Inf: returns: Inf, Inf

If obj(x) != Inf and ωf = Inf, returns: obj(x), Inf

MultiPrecisionR2.recomputeMu! — Method

recomputeMu!(m::FPMPNLPModel, solver::MPR2Solver{T,H}, πr::MPR2Precisions) where {T <: Tuple, H}

Recompute mu based on new precision levels. Performs only necessary operations to recompute mu. Possible operations are:

recompute candidate with higher prec FP format to decrease u

recompute ϕhat and ϕ with higher FP format for norm computation of x and s

recompute step with greater precision: decrease μ denominator

recompute gradient with higher precision to decrease ωg

recompute model reduction with higher precision to decrease αfunc(n,U[π.πΔ])

Does not make the over/underflow check as in main loop, since it is a repetition of the main loop with higher precisions and these issue shouldn't occur

Outputs:

g_recompute::Bool : true if gradient has been modified, false otherwise

See recomputeMuPrecSelection!

MultiPrecisionR2.recomputeMuPrecSelection! — Method

recomputeMuPrecSelection!(π::MPR2Precisions, πr::MPR2Precisions, πmax)

Default strategy to select new precisions to recompute μ in the case where μ > κₘ. Return false if no precision can be increased.

Modified arguments:

πr: contains new precision that will be used to recompute mu, see recomputeMu!

Ouptputs

max_prec::bool : return true if maximum precision levels have been reached

MultiPrecisionR2.recompute_g_default! — Method

recompute_g_default!(m::FPMPNLPModel, solver::MPR2Solver{T,H}, stats::GenericExecutionStats, e::E)

Increase operation precision levels until sufficiently small μ indicator is achieved. See also computeMu, recomputeMuPrecSelection!, recomputeMu!

Outputs:

::bool : returns false if couldn't reach sufficiently small evaluation error or overflow occured.

MultiPrecisionR2.reset! — Method

reset!(mpnlp::AbstractMPNLPModel)

Reset evaluation count and model data (if appropriate) in mpnlp.

MultiPrecisionR2.reset! — Method

reset!(counters::MPCounters)

Reset evaluation counters

MultiPrecisionR2.selectPic_default! — Method

selectPic_default!(π::MPR2Precisions)

Default strategy for selecting FP format of candidate for the next evaluation. Updates solver.π.πf⁺.

MultiPrecisionR2.selectPif! — Method

selectPif!(m::FPMPNLPModel, solver::MPR2Solver{T,H}, ωfBound::H)

Select a precision for objective evaluation for candidate based on predicted evaluation error. Evaluation is predicted as:

Relative error:

Predicted value of objective at c: f(c) ≈ f(x) - ΔTk

Relative error model: ωf(c) = |f(c)| * RelErr

Interval error:

Predicted value of objective at c: f(c) ≈ f(x) - ΔTk

Interval evaluation error is proportional to f(x)

Interval evaluation error depends linearly with unit-roundoff

Other: Lowest precision that does not cast candidate in a lower prec FP format and f(c) predicted does not overflow

Modified arguments:

solver.π.πf⁺: updated with FP format index chosen for objective evaluation

MultiPrecisionR2.sum_counters — Method

sum_counters(mpnlp)

Sum all counters of problem mpnlp except cons, jac, jprod and jtprod.

MultiPrecisionR2.sum_counters — Method

sum_counters(c::MPCounters)

Sum all counters of c except cons, jac, jprod and jtprod.

MultiPrecisionR2.umpt! — Method

umpt!(x::Tuple, y::Vector{S})

Update the elements of the multi precision containers x with the value y. Only the elements of x of FP formats with precision greater or equal than y are updated (avoid rounding error).

MultiPrecisionR2.update_struct! — Method

update_struct!(str,other_str)

Update the fields of str with the fields of other_str.

MultiPrecisionR2.γfunc_test_error_bound — Method

γfunc_test_error_bound(n::Int,eps::AbstractFloat,γfunc)

Tests if γfunc callback provides strictly less than 100% error for dot product error of vector of size the dimension of the problem and the lowest machine epsilon.

MultiPrecisionR2.γfunc_test_template — Method

γfunc_test_template(γfunc)

Tests if γfunc callback function is properly implemented. Expected template: γfunc(n::Int,u::Float) -> Float

NLPModels.neval_cons — Method

neval_cons(nlp)
neval_cons(nlp,T)

Get the total number (all FP formats) of cons evaluations. If extra argument T is provided, returns the number of cons evaluations for the given FP format T.

NLPModels.neval_cons_lin — Method

neval_cons_lin(nlp)
neval_cons_lin(nlp,T)

Get the total number (all FP formats) of cons evaluations. If extra argument T is provided, returns the number of cons evaluations for the given FP format T.

NLPModels.neval_cons_nln — Method

neval_cons_nln(nlp)
neval_cons_nln(nlp,T)

Get the total number (all FP formats) of cons evaluations. If extra argument T is provided, returns the number of cons evaluations for the given FP format T.

NLPModels.neval_grad — Method

neval_grad(nlp)
neval_grad(nlp,T)

Get the total number (all FP formats) of grad evaluations. If extra argument T is provided, returns the number of grad evaluations for the given FP format T.

NLPModels.neval_hess — Method

neval_hess(nlp)
neval_hess(nlp,T)

Get the total number (all FP formats) of hess evaluations. If extra argument T is provided, returns the number of hess evaluations for the given FP format T.

NLPModels.neval_hprod — Method

neval_hprod(nlp)
neval_hprod(nlp,T)

Get the total number (all FP formats) of hprod evaluations. If extra argument T is provided, returns the number of hprod evaluations for the given FP format T.

NLPModels.neval_jac — Method

neval_jac(nlp)
neval_jac(nlp,T)

Get the total number (all FP formats) of jac evaluations. If extra argument T is provided, returns the number of jac evaluations for the given FP format T.

NLPModels.neval_jac_lin — Method

neval_jac_lin(nlp)
neval_jac_lin(nlp,T)

Get the total number (all FP formats) of jac evaluations. If extra argument T is provided, returns the number of jac evaluations for the given FP format T.

NLPModels.neval_jac_nln — Method

neval_jac_nln(nlp)
neval_jac_nln(nlp,T)

Get the total number (all FP formats) of jac evaluations. If extra argument T is provided, returns the number of jac evaluations for the given FP format T.

NLPModels.neval_jcon — Method

neval_jcon(nlp)
neval_jcon(nlp,T)

Get the total number (all FP formats) of jcon evaluations. If extra argument T is provided, returns the number of jcon evaluations for the given FP format T.

NLPModels.neval_jgrad — Method

neval_jgrad(nlp)
neval_jgrad(nlp,T)

Get the total number (all FP formats) of jgrad evaluations. If extra argument T is provided, returns the number of jgrad evaluations for the given FP format T.

NLPModels.neval_jhess — Method

neval_jhess(nlp)
neval_jhess(nlp,T)

Get the total number (all FP formats) of jhess evaluations. If extra argument T is provided, returns the number of jhess evaluations for the given FP format T.

NLPModels.neval_jhprod — Method

neval_jhprod(nlp)
neval_jhprod(nlp,T)

Get the total number (all FP formats) of jhprod evaluations. If extra argument T is provided, returns the number of jhprod evaluations for the given FP format T.

NLPModels.neval_jprod — Method

neval_jprod(nlp)
neval_jprod(nlp,T)

Get the total number (all FP formats) of jprod evaluations. If extra argument T is provided, returns the number of jprod evaluations for the given FP format T.

NLPModels.neval_jprod_lin — Method

neval_jprod_lin(nlp)
neval_jprod_lin(nlp,T)

Get the total number (all FP formats) of jprod evaluations. If extra argument T is provided, returns the number of jprod evaluations for the given FP format T.

NLPModels.neval_jprod_nln — Method

neval_jprod_nln(nlp)
neval_jprod_nln(nlp,T)

Get the total number (all FP formats) of jprod evaluations. If extra argument T is provided, returns the number of jprod evaluations for the given FP format T.

NLPModels.neval_jtprod — Method

neval_jtprod(nlp)
neval_jtprod(nlp,T)

Get the total number (all FP formats) of jtprod evaluations. If extra argument T is provided, returns the number of jtprod evaluations for the given FP format T.

NLPModels.neval_jtprod_lin — Method

neval_jtprod_lin(nlp)
neval_jtprod_lin(nlp,T)

Get the total number (all FP formats) of jtprod evaluations. If extra argument T is provided, returns the number of jtprod evaluations for the given FP format T.

NLPModels.neval_jtprod_nln — Method

neval_jtprod_nln(nlp)
neval_jtprod_nln(nlp,T)

Get the total number (all FP formats) of jtprod evaluations. If extra argument T is provided, returns the number of jtprod evaluations for the given FP format T.

NLPModels.neval_obj — Method

neval_obj(nlp)
neval_obj(nlp,T)

Get the total number (all FP formats) of obj evaluations. If extra argument T is provided, returns the number of obj evaluations for the given FP format T.